Fixed Infinite loop in the questionBank.txt parsing
This was caused by how the question bank was finding the end of a question block.
Right now it is looking for any line that has under 1 elements, which will definitely
work if the questionBank.txt is formatted exactly how it should be, but we are all human
so simple things like a space character can slip past us (As in this case).

I left the parsing code the same for now but will make changes in a bit to instead
maybe just search that string for a $ character.